fix: fix artifacts from merge with staging
