+++ tests/features/steps/well-notes.py
@@ -33,24 +33,6 @@ def step_impl(context):
         assert note, f"{k} Note is empty"
 
 
-@when("the user retrieves the well by ID via path parameter")
-def step_impl(context):
-    context.response = context.client.get(
-        f"thing/water-well/{context.objects['wells'][0].id}"
-    )
-    context.notes = {}
-
-
-@then(
-    "null values in the response should be represented as JSON null (not placeholder strings)"
-)
-def step_impl(context):
-    data = context.response.json()
-    for k, v in data.items():
-        if v == "":
-            assert v is None, f"Value for key {k} is an empty string but should be null"
-
-
 @then(
     "the response should include location notes (i.e. driving directions and geographic well location notes)"
 )
